pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Initial Setup"
authors = ["icarus77"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11"
fastapi = { extras = ["all"], version = ">=0.115.12,<0.116.0" }

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"
pre-commit = "^4.2.0"
colorlog = "^6.9.0"
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
setuptools = "^80.7.1"
prisma = "^0.15.0"
pydantic-settings = "^2.9.1"
ruff = "^0.11.9"
uvicorn = "^0.34.2"
firebase-admin = "^6.8.0"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
skip_glob = ["src/prisma_client/*"]

[tool.ruff]
line-length = 88

[tool.mypy]
strict = true
exclude = ["src/prisma_client", "src/prisma_client/*"]

[[tool.mypy.overrides]]
module = "src.prisma_client.*"
ignore_errors = true
follow_imports = "skip"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.mypy.overrides]]
module = ["firebase_admin", "firebase_admin.*"]
follow_untyped_imports = false
ignore_missing_imports = true
